+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+# 1.0.0.1 - 2024-06-16
 - support fetch the value from UHUGEINT type column.
 - add `DuckDB::Appender#append_uhugeint`.
 - DuckDB::Result supports ARRAY column type (only when DuckDB::Result.use_chunk_each is true).